improvements

+++ contracts/sales/INFTSales.sol
@@ -3,7 +3,7 @@ pragma solidity 0.8.11;
 
 interface INFTSales {
     function initialize(uint64 seriesId, address currency, uint256 price, address beneficiary, uint192 autoindex, uint64 duration, uint32 rateInterval, uint16 rateAmount) external;
-    function specialPurchaseByLicenses(uint256 amount, address[] memory accounts, address[] memory contracts, uint32[] memory tokenIds) external payable;
+    function specialPurchaseByLicenses(uint256 amount, address[] memory accounts, address[] memory contracts, uint256[] memory tokenIds) external payable;
     function specialPurchase(uint256 amount, address[] memory accounts) external payable;
     function purchase(uint256 amount, address[] memory accounts) external payable;
     function remainingDays(uint256 tokenId) external view returns (uint64);
